+++ src/App.vue
@@ -137,4 +137,8 @@
     -moz-box-sizing: border-box;
     box-sizing: border-box;
   }
+
+  .alert-container {
+    z-index: 9999;
+  }
 </style>