+++ src/store/appType/ecommerce/state.js
@@ -1,0 +1,13 @@
+export default {
+  loadingEcommerceApps: true,
+  errorEcommerceApps: null,
+  ecommerceAppsList: null,
+
+  loadingConnectVtexCatalog: false,
+  errorConnectVtexCatalog: null,
+  connectVtexCatalogData: null,
+
+  loadingVtexAppUuid: false,
+  errorVtexAppUuid: null,
+  generatedVtexAppUuid: null,
+};