+++ src/store/myApps/mutations.js
@@ -1,8 +1,11 @@
 export default {
-  GET_CONFIGURED_APPS_REQUEST(state) {
-    state.loadingConfiguredApps = true;
+  GET_CONFIGURED_APPS_REQUEST(state, skipLoading = false) {
+    state.loadingConfiguredApps = !skipLoading;
     state.errorConfiguredApps = null;
-    state.configuredApps = null;
+
+    if (!skipLoading) {
+      state.configuredApps = null;
+    }
   },
   GET_CONFIGURED_APPS_SUCCESS(state, data) {
     state.configuredApps = data;